Added FSM
Added the FSM, but it doesn't work right now. The roam works, but not how I want it to. Fleeing currently doesn't work. I had the change the previous random picker as it caused all the cows to group up. Also not sure why, but the score system doesn't work anymore.